Fixes following issues
1. A query containing GRAPH_TABLE reference used in a UDF causes segmentation
fault: Earlier version of patches used p_post_columnref_hook to transform a
property reference in a GRAPH_TABLE reference. But related hooks and members in
ParseState are used by other non-core modules like plpgsql. If a query inside a
plpgsql function has GRAPH_TABLE reference in it the usage of these hooks
conflicts. Since GRAPH_TABLE is a core feature, it's better to introduce a new
member ParseState::p_graph_table_pstate, of type GraphTableParseState, to hold
the namespace for the GRAPH_TABLE reference being transformed. It is inline
with ParseState::p_parent_cte or ParseState::p_queryEnv which have similar
purposes. With that change definition of GraphTableParseState is moved to
parsenodes.h where CommonTableExpr and QueryEnvironment are also defined. The
function using the ParseState::p_graph_table_pstate is renamed to
transformGraphTablePropertyRef() to be inline with names of other similar
functions. This change also allows GraphTableParseState argument to be removed
from a few functions.

2. ERROR:  unrecognized lock mode: 0 via ScanQueryForLocks(): After fixing the
first bug, I ran into above error. An RTE_GRAPH_TABLE is converted to
RTE_SUBQUERY after transformation with relid = OID of the property graph being
referenced in GRAPH_TABLE clause. Functions called from ScanQueryForLocks()
take locks when RangeTblRef::relid is valid; which is desirable since we want
to lock the property graph while it's being used. So set
RangeTblRef::rellockmode to AccessShareLock.

3. Prohibits subquery within GRAPH_TABLE reference: In order to support it the
hasSublinks flag needs to be transferred to the queries produced as a result of
rewriting graph table RTE. It needs a bit of code and some non-trivial testing.
So not done right now.

4. Support edge patterns in any direction: Such edge patterns satisfy edges in
either direction.

5. If there are multiple foreign keys on an edge table, the last one that appears
in the list returned by RelationGetFKeyList() is used to create a vertex
reference from that edge. This may lead to wrong edge-vertex linkage. Remember
the correct one instead.

Additional changes
1. Test for a join between GRAPH_TABLE and a regular table.
2. Changed a few loops over lists to use foreach_node reducing and simplifying
code a bit.
3. Improved comments.

Author: Ashutosh Bapat
Reported By: Ajay Pal, Junwang Zhou