WIP: Do not print empty columns table for a property graph
A property graph does not have any columns by itself. Thus it doesn't
need the table describing columns in \d[+] output.

The patch needs following work
1. Move the code to collect column information in a separate function,
   which can be skipped for property graphs. This avoids one indentation
   level and makes describeOneTableDetails() easy to follow.
2. Even though the borders are suppressed there is an extra blank line
   in \d output. Avoid that. It requires some adjustment to
   printTableOpt or some minion of printTable to avoid printing
   everything related to column description table.

   OR

If we decide that we want to output a different description (like
sequences) we need to decide what that description should be and then
implement the same.

Ashutosh Bapat